refactor dashboard config
